Other resolutions now fully supported

+++ fractal-cpp/src/cs.shader
@@ -22,7 +22,7 @@ layout(std430, binding = 1) buffer bufferIn
 const vec4 notEscapedColor = vec4(1.0, 0.0, 0.0, 0.0);
 
 /* Width and height of the window in pixels. */
-uniform vec2 screenResolution;
+uniform ivec2 screenResolution;
 /* The zoom factor by which everything is scaled. */
 uniform double zoom;
